Fix: change the PatternedPairGenerator algorithm
The new algorithms mimics randomness better. Especially, the
(nbElements/2) quantity in PatternedPairGenerator.GetPair can be set to
any number. It determines the number of pairs between to occurences of a
particulat index. For example, if nbElements/2 is 3, and the 20th pair
is (1,4), you have the guarantee that 1 and 4 won't appear in the 3 next
pairs.